simplify example providers
