--- examples/issuer/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "issuer"
-publish = false
-edition.workspace = true
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow.workspace = true
-axum.workspace = true
-axum-extra.workspace = true
-base64ct.workspace = true
-chrono.workspace = true
-credibil-did.workspace = true
-credibil-infosec.workspace = true
-credibil-vc = { workspace = true, features = ["issuer"] }
-ed25519-dalek = "2.1.1"
-oauth2 = { version = "5.0.0" }
-rand_core = "0.6.4"
-serde.workspace = true
-serde_json.workspace = true
-tokio.workspace = true
-tower-http.workspace = true
-tracing.workspace = true
-tracing-subscriber.workspace = true
-uuid.workspace = true
