--- tests/utils/issuer/data/client.json
@@ -1,16 +1,0 @@
-{
-    "client_id": "96bfb9cb-0513-7d64-5532-bed74c48f9ab",
-    "client_name": "Wallet",
-    "redirect_uris": [
-        "http://localhost:3000/callback"
-    ],
-    "grant_types": [
-        "authorization_code",
-        "urn:ietf:params:oauth:grant-type:pre-authorized_code"
-    ],
-    "response_types": [
-        "code"
-    ],
-    "scope": "openid credential",
-    "credential_offer_endpoint": "openid-credential-offer://"
-}
