--- tests/utils/issuer/data/datasets.json
@@ -1,67 +1,0 @@
-{
-    "normal_user": {
-        "PHLEmployeeID": {
-            "configuration_id": "EmployeeID_W3C_VC",
-            "claims": {
-                "given_name": "Normal",
-                "family_name": "Person",
-                "email": "normal.user@example.com",
-                "address": {
-                    "street_address": "123 Fake St",
-                    "locality": "Wellington"
-                }
-            }
-        },
-        "PHLDeveloper": {
-            "configuration_id": "Developer_W3C_VC",
-            "claims": {
-                "name": "A. Developer",
-                "role": "Lead",
-                "proficiency": "3"
-            }
-        },
-        "OpenSourceDeveloper": {
-            "configuration_id": "Developer_W3C_VC",
-            "claims": {
-                "name": "A. Mazing",
-                "role": "Hacker",
-                "proficiency": "2"
-            }
-        },
-        "DriverLicence": {
-            "configuration_id": "org.iso.18013.5.1.mDL",
-            "claims": {
-                "given_name": "Normal",
-                "family_name": "Person",
-                "email": "normal.user@example.com"
-            }
-        },
-        "Identity": {
-            "configuration_id": "Identity_SD_JWT",
-            "claims": {
-                "given_name": "Normal",
-                "family_name": "Person",
-                "email": "normal.user@example.com",
-                "address": {
-                    "street_address": "123 Fake St",
-                    "locality": "Wellington"
-                }
-            }
-        }
-    },
-    "pending_user": {
-        "PHLEmployeeID": {
-            "configuration_id": "EmployeeID_W3C_VC",
-            "claims": {
-                "given_name": "Pending",
-                "family_name": "Person",
-                "email": "pending.user@example.com",
-                "address": {
-                    "street_address": "123 Fake St",
-                    "locality": "Wellington"
-                }
-            },
-            "pending": true
-        }
-    }
-}
