--- tests/utils/verifier/data/verifier.json
@@ -1,21 +1,0 @@
-{
-    "client_id": "http://localhost:8080",
-    "client_name": "Verifier",
-    "redirect_uris": [
-        "http://localhost:3000/callback"
-    ],
-    "response_types": [
-        "vp_token",
-        "id_token vp_token"
-    ],
-    "vp_formats": {
-        "jwt_vp_json": {
-            "alg": [
-                "ES256K"
-            ],
-            "proof_type": [
-                "JsonWebSignature2020"
-            ]
-        }
-    }
-}
