+++ tests/vci_deferred.rs
@@ -1,5 +1,7 @@
 //! Deferred Issuance Tests
 
+#[path = "../examples/issuer/provider/mod.rs"]
+mod provider;
 mod utils;
 
 use std::sync::LazyLock;
@@ -13,7 +15,7 @@ use credibil_vc::oid4vci::types::{
 };
 use credibil_vc::oid4vci::{JwtType, endpoint};
 use insta::assert_yaml_snapshot as assert_snapshot;
-use utils::issuer::{CREDENTIAL_ISSUER as ALICE_ISSUER, PENDING_USER, ProviderImpl};
+use provider::{CREDENTIAL_ISSUER as ALICE_ISSUER, PENDING_USER, ProviderImpl};
 use utils::wallet::{self, Keyring};
 
 static BOB_KEYRING: LazyLock<Keyring> = LazyLock::new(wallet::keyring);
