+++ tests/vci_issuance.rs
@@ -2,6 +2,8 @@
 
 //! Pre-Authorized Code Flow Tests
 
+#[path = "../examples/issuer/provider/mod.rs"]
+mod provider;
 mod utils;
 
 use std::sync::LazyLock;
@@ -17,7 +19,7 @@ use credibil_vc::oid4vci::{JwtType, endpoint};
 use credibil_vc::sd_jwt::SdJwtClaims;
 use insta::assert_yaml_snapshot as assert_snapshot;
 use sha2::{Digest, Sha256};
-use utils::issuer::{CREDENTIAL_ISSUER as ALICE_ISSUER, NORMAL_USER, ProviderImpl};
+use provider::{CREDENTIAL_ISSUER as ALICE_ISSUER, NORMAL_USER, ProviderImpl};
 use utils::wallet::{self, Keyring};
 
 static BOB_KEYRING: LazyLock<Keyring> = LazyLock::new(wallet::keyring);
